Extract blob-safe MIME types
